Binarization fixes, k_factor default 0.05, and quality consolidation
Fable-review bugfixes (all bugs found by Fable code review):
- #1: Bottom-edge region bounds in rgb_to_grayscale_simd_direct.
      y_end * source_width → (y_end - 1) * source_width so
      full-page crops (y_end == page_height) don't silently blank out.
- #2: GPU sum-of-squares SAT from f32 → wrapping u32 across all 4 WGSL
      shaders (build_padded, integral_rows/cols, final) and Rust host.
      f32 prefix sum loses ULP precision at page scale (ULP ~4096 @ 7e10),
      causing position-dependent threshold jitter. u32 SAT is exact:
      max window Σg² = 101² × 255² ≈ 6.6e8 < 2³².
- #3: GPU Otsu computed on bg-normalized histogram instead of raw gray.
      Premask 255 spike in layout mode caused raw-Otsu to lock onto
      paper|masked-white split, blackening the page. Normalizing first
      collapses the spike onto the paper mode, matching CPU semantics.
      Re-enables GPU binarization in crop-free and layout-detection modes
      (reverts disable_gpu workarounds in policies.rs, pnginference.rs).
- #5: compute_adaptive_gpu_constants now uses the same linearized BT.709
      luma as the shader thresholds (was 601-gamma approximation).
- #8: Heavy-path region extraction: copy_from_slice with hard bounds error
      instead of pixel-by-pixel copy with silent zero-fill.
- #9: Clustered-screen dither in linear-reflectance space via SRGB_GRAY_TO_
      LINEAR_U8 LUT. sRGB 128 ≈ 21% luminance was rendering as ~50% dot
      coverage, making midtones appear too light.
- #10: DjVu background subsample (c44 -bgsubsample). Encode BG44 at 1/N
      (default 3×); viewers upsample. Biggest size lever for DjVu output.
      Validated via djvulibrust/examples/bg_subsample_check.rs.
- #11: quality_policy module — all scattered quality constants consolidated
      behind named intents (full_page_jp2, region_jpeg, etc.).
- Smaller: percentile_from_hist256 replaces sorted-plane percentiles in
  normalize_crop_binarization_input (O(n) vs O(n log n)).

k_factor default change (parameter-sweep validated):
- Default Sauvola k_factor reduced from 0.20 to 0.05 after sweep testing
  on clean-white (wheels) and yellowed (sahib) all-text pages. At 0.20,
  Otsu arm collapses on near-unimodal clean scans, taking chunks out of
  thin strokes. k=0.05 fused (debug_mode=0) produces clean text on both
  light and dark backgrounds without chunking.
- DEFAULT_K_FACTOR const defined in legencode/src/types.rs as single source
  of truth, re-exported for GUI (lege::DEFAULT_K_FACTOR) and lege-ocr.
  lege-gpu duplicates with doc link (circular dep constraint).

Revived binarization tooling:
- binarize_compare.rs: per-arm CPU-vs-GPU dump wired to GPU debug modes
  0–6 (fused, Sauvola, Otsu, bg, normalized, mean, stddev).
- binarize_sweep.rs: k-factor/parameter sweep harness.

+++ lege-gpu/src/binarization/wgpu/shaders/adaptive_build_padded.wgsl
@@ -8,7 +8,7 @@ struct BinarizeParams {
 @group(0) @binding(0) var<uniform> params: BinarizeParams;
 @group(0) @binding(1) var<storage, read> gray_src: array<u32>; // 1 u32 per pixel (sRGB gray from linearize pass)
 @group(0) @binding(2) var<storage, read_write> padded_gray: array<u32>;
-@group(0) @binding(3) var<storage, read_write> padded_sq: array<f32>;
+@group(0) @binding(3) var<storage, read_write> padded_sq: array<u32>;
 
 fn reflect_101(idx: i32, len: i32) -> i32 {
     if len <= 1 { return 0; }
@@ -33,5 +33,8 @@ fn main(@builtin(global_invocation_id) id: vec3<u32>) {
     let g = gray_src[src_idx] & 255u;
     let out_idx = py * params.padded_width + px;
     padded_gray[out_idx] = g;
-    padded_sq[out_idx] = f32(g) * f32(g);
+    // Second moment as exact u32. Max window sum-of-squares is
+    // 101^2 * 255^2 ~= 6.6e8 < 2^32, so a wrapping-u32 SAT gives exact window
+    // sums at any image size (an f32 prefix sum loses ULP precision at page scale).
+    padded_sq[out_idx] = g * g;
 }
